added front end

+++ index.js
@@ -9,28 +9,37 @@ const ffmpeg = require("ffmpeg-static");
 const app = express();
 const axios = require('axios');
 const bodyParser = require("body-parser");
-app.use(bodyParser.urlencoded({ extended: false }));
 const port = 80;
 const rick = 'https://www.youtube.com/watch?v=dQw4w9WgXcQ&ab_channel=RickAstley';
+app.use(bodyParser.urlencoded({ extended: false }));
 app.set('trust proxy', true)
 app.use(express.static('public'));
 app.set("view engine", "ejs");
 
-async function del() {
+function del() {
     glob("download/*.mkv", {}, function (er, files) {
         files.forEach(function (x) {
             console.log(x);
-            fs.unlinkSync(x);
+            let id = x.substring(x.length-15, x.length-4)
+            let status = `status/${id}.status`
+            axios.get(`/status?v=${id}`)
+                .then(response=> {
+                    if(response.data.status === 'done'){
+                        fs.unlinkSync(x);
+                        fs.unlinkSync(status);
+                    }
+                });
         });
-    })
+    });
 }
 
 setInterval(function (){
     del()
-}, 3600000)
+}, 1800000)
 
 
 function down(ref, res, file_path) {
+    fs.writeFile(`${__dirname}/status/${ref}.status`, "",function (){})
     try {
         const audio = ytdl(ref, {quality: 'highestaudio'})
         const video = ytdl(ref, {quality: 'highestvideo'})
@@ -53,7 +62,13 @@ function down(ref, res, file_path) {
             ],
         });
         ffmpegProcess.on('close', () => {
-            res.redirect(`/download?v=${ref}`)
+            //res.redirect(`/download?v=${ref}`)
+            fs.writeFile(`${__dirname}/status/${ref}.status`, "done", err => {
+                if (err) {
+                    console.error(err);
+                }
+                // file written successfully
+            });
         });
         ffmpegProcess.on('error', function(err) {
             console.log(err)
@@ -64,6 +79,7 @@ function down(ref, res, file_path) {
         console.log(e);
     }
 }
+
 app.post("/login", (req, res) => {
     const { name, password } = req.body;
 
@@ -85,12 +101,12 @@ app.get('/watch', function (req, res) {
             let exists = false;
             let vid_path = `${video_id}.mkv`
             let file_path = `${__dirname}/download/${vid_path}`;
-            try{
+            try{ //nominal entry
                 axios.get(`https://api.unblockvideos.com/youtube_restrictions?id=${video_id}`)
                     .then(r=> {
                         let cont = true;
                         const dat = r.data[0].blocked;
-                        //console.log(dat);
+
                         if(dat.includes("US")){
                             console.log("Not Available for download");
                             res.render("error",{
@@ -98,6 +114,7 @@ app.get('/watch', function (req, res) {
                             })
                             cont = false;
                         }
+
                         try {
                             if (fs.existsSync(file_path)) {
                                 exists = true;
@@ -105,15 +122,22 @@ app.get('/watch', function (req, res) {
                         } catch (err) {
                             exists = false;
                         }
-                        //console.log(video_id);
                         if (!exists && cont) {
+                            res.render(`watch`,{
+                                link:video_id
+                            })
                             down(video_id, res, file_path);
                         }
-                        else if(exists ) {
-                            res.redirect(`/download?v=${queryObject.v}`);
+                        else if(exists) {
+                            //res.redirect(`/download?v=${queryObject.v}`);
+                            res.render(`watch`,{
+                                link:video_id
+                            })
                         }
                         else if(!cont){
-                            //pass
+                            res.render("error",{
+                                error:"Other Error"
+                            })
                         }
                     })
             }
@@ -122,18 +146,14 @@ app.get('/watch', function (req, res) {
             }
 
         } else {
-            //res.redirect(rick);
-            //console.log(`Rick Rolled${req.ip}`);
             res.render("error",{
                 error:"Invalid ID"
             })
         }
     } else if (queryObject.v === undefined) {
-        //res.redirect(rick);
         res.render("error",{
             error:"No Query Specified"
         })
-        //console.log(`Rick Rolled ${req.ip}`);
     }
 
 })
@@ -141,10 +161,37 @@ app.get('/watch', function (req, res) {
 app.get('/rick', function (req, res) {
     res.redirect(rick);
 })
+
 app.get('/', function (req, res) {
     //res.redirect(rick);
     res.render("index");
 })
+
+app.get('/status',function (req, res) {
+    const queryObject = url.parse(req.url, true).query;
+    let video_id = queryObject.v;
+    if (video_id != null) {
+        if (muxer.validateID(`${video_id}`)) {
+            fs.readFile(`${__dirname}/status/${video_id}.status`, 'utf8', (err, data) => {
+                if (err) {
+                    //console.error(err);
+                    res.status(500)
+                    return;
+                }
+                let re = {
+                    "status":data
+                };
+                res.status(200).json(re);
+            });
+
+        }
+        else{
+            res.status(403)
+        }
+    }
+})
+
+
 app.get('/download', function (req, res) {
     const queryObject = url.parse(req.url, true).query;
     if (queryObject.v != null) {
